+++ src/pages/MainPage.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+import "../css/MainPage.css";
+import Main from "../component/mainpage/Main";
+
+export default function MainPage() {
+  return <Main />;
+}
