+++ src/pages/SignUpPage.jsx
@@ -1,0 +1,6 @@
+import React from "react";
+import SignUp from "../component/sign/SignUp";
+
+export default function SignUpPage() {
+  return <SignUp />;
+}
