+++ src/pages/SigninPage.jsx
@@ -1,0 +1,6 @@
+import React from "react";
+import SignIn from "../component/sign/SignIn";
+
+export default function SigninPage() {
+  return <SignIn />;
+}
